Some Fixes in list and schedule

+++ lib/app/controllers/schedule_controller.dart
@@ -1,11 +1,32 @@
 import 'package:get/get.dart';
+import 'package:tfg_app/app/mappers/subject_model_mapper.dart';
+import 'package:tfg_app/data/model/classroom_model.dart';
 import 'package:tfg_app/data/model/data_model.dart';
+import 'package:tfg_app/data/model/degree_model.dart';
+import 'package:tfg_app/data/model/department_model.dart';
 import 'package:tfg_app/data/model/subject_model.dart';
 import 'package:tfg_app/domain/data_repository.dart';
 
 class ScheduleController extends GetxController with StateMixin<DataModel> {
   final DataRepository dataRepository;
   var subjects = Rx<List<SubjectModel>>([]);
+  Rx<SubjectModel> selectedSubject = SubjectModel(
+      id: -1,
+      name: "",
+      acronym: "",
+      classGroup: "",
+      seminary: false,
+      laboratory: false,
+      english: false,
+      time: 0,
+      semester: 0,
+      days: "",
+      hours: "",
+      turns: "",
+      classroom: ClassroomModel(id: -1, name: "", pavilion: "", acronym: ""),
+      department: DepartmentModel(id: -1, name: "", acronym: ""),
+      degree: DegreeModel(id: -1, name: "", num_semesters: 0, year: ""),
+      color: 0).obs;
   var morning5Rows = List.generate(
           12, (i) => List<SubjectModel?>.filled(5, null, growable: false),
           growable: false)
@@ -29,7 +50,6 @@ class ScheduleController extends GetxController with StateMixin<DataModel> {
   void onInit() {
     getData();
     super.onInit();
-    print(morning5Rows);
   }
 
   void getData() {
@@ -47,20 +67,79 @@ class ScheduleController extends GetxController with StateMixin<DataModel> {
     if (morning) {
       SubjectModel? item = subjects.value.firstWhereOrNull(
           (element) => element.id == morning5Rows.value[column][row]?.id);
-      item ??= morning5Rows.value[column][row];
-      var pos = subjects.value.indexOf(item!);
-      var newItem = item.deleteItem(newTime: item.time + 30);
-      subjects.value[pos] = newItem;
+      if(item != null) {
+        var pos = subjects.value.indexOf(item);
+        var newItem = item.deleteItem(newTime: item.time + 30);
+        subjects.value[pos] = newItem;
+      } else {
+        var newSubject = selectedSubject.value.copyWith(newTime: selectedSubject.value.time + 30);
+        subjects.value.add(newSubject);
+      }
       morning5Rows.value[column][row] = null;
     } else {
       SubjectModel? item = subjects.value.firstWhereOrNull(
           (element) => element.id == afternoon5Rows.value[column][row]?.id);
-      item ??= afternoon5Rows.value[column][row];
-      var pos = subjects.value.indexOf(item!);
-      var newItem = item.deleteItem(newTime: item.time + 30);
-      subjects.value[pos] = newItem;
+      if(item != null) {
+        var pos = subjects.value.indexOf(item);
+        var newItem = item.deleteItem(newTime: item.time + 30);
+        subjects.value[pos] = newItem;
+      } else {
+        var newSubject = selectedSubject.value.copyWith(newTime: selectedSubject.value.time + 30);
+        subjects.value.add(newSubject);
+      }
       afternoon5Rows.value[column][row] = null;
     }
     update();
+  }
+
+  void dragItemSuccessfully(int index) {
+    selectedSubject = subjects.value
+        .firstWhere((element) =>
+            subjects.value
+                .map((data) => data.toSubjectBox())
+                .toList()[index]
+                .subject
+                .id ==
+            element.id)
+        .obs;
+    var pos = subjects.value.indexOf(selectedSubject.value);
+    if (subjects.value[pos].time == 0) {
+      subjects.value.removeAt(pos);
+    }
+    update();
+  }
+
+  void completeDrag(SubjectModel subject, bool morning, int row, int column) {
+    morning
+        ? assignItemMorning(subject, row, column)
+        : assignItemAfternoon(subject, row, column);
+  }
+
+  void assignItemMorning(SubjectModel subject, int row, int column) {
+    if (morning5Rows.value[column][row] == null) {
+      morning5Rows.value[column][row] = subject;
+      var pos = subjects.value.indexOf(selectedSubject.value);
+      subjects.value[pos] = selectedSubject.value
+          .copyWith(newTime: selectedSubject.value.time - 30);
+    }
+  }
+
+  void assignItemAfternoon(SubjectModel subject, int row, int column) {
+    if (afternoon5Rows.value[column][row] == null) {
+      afternoon5Rows.value[column][row] = subject;
+      var pos = subjects.value.indexOf(selectedSubject.value);
+      subjects.value[pos] = selectedSubject.value
+          .copyWith(newTime: selectedSubject.value.time - 30);
+    }
+  }
+
+  void startDrag(int index) {
+    selectedSubject.value = subjects.value.firstWhere((element) =>
+        subjects.value
+            .map((data) => data.toSubjectBox())
+            .toList()[index]
+            .subject
+            .id ==
+        element.id);
   }
 }
